StringWrapper

Provides a lazy abstract type StringWrapper <: AbstractString end interface implementation, delegating AbstractString API methods

index boundaries and sizes: firstindex, lastindex, length, ncodeunits, sizeof

getting elements getindex, iterate, codeunit

codeunits and indices: isvalid, thisind, prevind, nextind

SubString

LazyStrings.jl provides similar functionality for AbstractStrings as LazyArrays.jl for Vectors.

These sources have been reviewed for AbstractString interface methods.

https://docs.julialang.org/en/v1/base/strings/

What is the Interface of AbstractString?

I did not find any official table of API functions – please add methods in a PR if the API is missing what you need.

Note:

*, ^, repeat is not provided

findnext(f, ::AbstractString, ::Int) ignores specialised nextind #26202, is not tested

The package is used in CombinedParsers.jl for lookbehind parsers and parsers on a lazyly transformed String (e.g. lowercase). ReversedStrings.jl was the deprecated and moved into this package.

Custom StringWrapper

All you need is provide a LazyStrings.representation method.

LazyStrings.StringWrapper — Type

abstract type StringWrapper <: AbstractString end

Provides default sw::StringWrapper<:AbstractString API methods defering to representation(sw) call, by default sw.x.

Lazy CharMappedString

LazyStrings.CharMappedString — Type

CharMappedString(f::Function,x) <: AbstractString

String implementation lazily transforming characters. Used for parsing with MappedSequenceParser.

julia> using LazyStrings

julia> using BenchmarkTools

julia> @btime lmap(lowercase,"JuliaCon")[1]
  6.706 ns (0 allocations: 0 bytes)
'j': ASCII/Unicode U+006A (category Ll: Letter, lowercase)

julia> @btime map(lowercase,"JuliaCon")[1]
  93.681 ns (3 allocations: 144 bytes)
'j': ASCII/Unicode U+006A (category Ll: Letter, lowercase)

Lazy ReversedString

julia> using LazyStrings

julia> using BenchmarkTools

julia> @btime reverse("JuliaCon")
  28.694 ns (1 allocation: 32 bytes)
"noCailuJ"

julia> @btime reversed("JuliaCon")
  3.167 ns (0 allocations: 0 bytes)
"noCailuJ"

julia> @btime reverse(reverse("JuliaCon"))
  58.530 ns (2 allocations: 64 bytes)
"JuliaCon"

julia> @btime reversed(reversed("JuliaCon"))
  3.734 ns (0 allocations: 0 bytes)
"JuliaCon"

LazyStrings.ReversedString — Type

ReversedString{V}<:AbstractString

Lazy reversed V struct holding representation and caching lastindex.

LazyStrings.reversed — Function

reversed(x::AbstractString)

A lazy implementation of Base.reverse for Strings.

LazyStrings.reverse_index — Function

reverse_index(x::ReversedString,i)

Return corresponding index in unreversed String x.lastindex-i+1.

reverse_index(x::AbstractString,i)

Return original index i.

Base.firstindex — Function

Base.firstindex(x::ReversedString)

1

Base.lastindex — Function

Base.lastindex(x::ReversedString)

cached x.lastindex.

Base.length — Function

Base.length(x::ReversedString)

length(x.representation)

Base.ncodeunits — Function

Base.ncodeunits(x::ReversedString)

ncodeunits(x.representation)

Base.thisind — Function

Base.thisind(x::ReversedString,i::Integer)

reverse_index(thisind(x.representation,reverse_index(x,i))).

Base.prevind — Function

Base.prevind(x::ReversedString,i[,n=1])

reverse_index(prevind(x.representation,reverse_index(x,i))[,n]).

Base.nextind — Function

Base.nextind(x::ReversedString,i[,n=1])

reverse_index(nextind(x.representation,reverse_index(x,i))[,n]).

Base.iterate — Function

Base.iterate(x::ReversedString[, state])

Iterate lazy reversed string.

Base.getindex — Function

Base.getindex(x::ReversedString,is::UnitRange{<:Integer})

ReversedString of getindex on representation.

Note

getindex is creating a new String representation. See SubString

Base.getindex(x::ReversedString,i::Integer)

getindex(x.representation,reverse_index(x,i)).

Base.codeunit — Function

Base.ncodeunits(x::ReversedString)

codeunit(x.representation,reverse_index(x,i)).

Base.isvalid — Function

Base.isvalid(x::ReversedString,i::Int)

isvalid(x.representation,reverse_index(x,i)).

Base.SubString — Type

Base.SubString(x::ReversedString,start::Int,stop::Int)

Create a reversed SubString using reverse_index.